补水仪电量

+++ OznerLibrary/WaterReplenishmentMeter/WaterReplenishmentMeterStatus.h
@@ -13,7 +13,7 @@
 
 -(instancetype)init;
 @property (readonly) BOOL power;
-@property (readonly) int battery;
+@property (readonly) float battery;
 @property (readonly) BOOL testing;
 @property (readonly) float moisture;
 @property (readonly) float oil;
